+++ src/components/Header/index.tsx
@@ -1,57 +1,17 @@
-import { useEffect, useState } from 'react'
-import { FiMenu } from 'react-icons/fi'
+import { Navbar } from '../Navbar'
 
 import {
   Container,
-  Nav, 
-  NavContent, 
-  Logo, 
-  Hamburguer, 
-  Menu, 
-  MenuLink,
   HeaderContainer,
   HeaderContent,
   Button
 } from './styles'
 
 export function Header() {
-  const [hamburguerIsOpen, setHamburguerIsOpen] = useState(false)
-  const [navBackground, setnavBackground] = useState('backgroundTransparent')
-
-  useEffect(function mount() {
-    function handleScroll() {
-      if(window.scrollY >= 90) {
-        setnavBackground('backgroundSolid')
-      } else {
-        setnavBackground('backgroundTransparent')
-      }
-    }
-    window.addEventListener('scroll', handleScroll)
-
-    return function unMount() {
-      window.removeEventListener('scroll', handleScroll)
-    }
-  })
-
   return (
     <Container>
-      <Nav navBackground={navBackground}>
-        <NavContent>
-          <Logo href="/">
-            <h1>
-              SaganSat
-            </h1>
-          </Logo>
-          <Hamburguer onClick={() => setHamburguerIsOpen(!hamburguerIsOpen)}>
-            <FiMenu size={30} />
-          </Hamburguer>
-          <Menu isOpen={hamburguerIsOpen}>
-            <MenuLink href="/">Sobre a equipe</MenuLink>
-            <MenuLink href="/">Sobre a pesquisa</MenuLink>
-            <MenuLink href="/">Como você pode ajudar?</MenuLink>
-          </Menu>
-        </NavContent>
-      </Nav>
+      <Navbar />
+      
       <HeaderContainer>
         <HeaderContent>
           <h1>SAGANSAT-EA</h1>
